requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
numpy
